Add RLP encoding subcommand for ibftExtraData to be inserted in genesis file
This is intentionally not a generic command as anyway RPL encoding is specific
to each type of data.
It was coded with efficiency in mind and as there's only one encodable type for
the moment, the --type option is not provided. We'll add the option later if
needed and the current type will be the default one.
However a RLPEncodable interface is added to handle the RLP encode behaviour.
JSON annotations were added for type mapping.

made the file output optional and write to standard output by default, same for
standard input.

Some mnore tests are required, this is a preview version.