Updated code to use an adapter instead of modifying the deep classes with JSON @
Also found a way to make the enum able to return a type in a generic way using
an interface so the `--type` option can be used. We still have only one type
but at least it's ready for two.

Still some unit tests to add in the next commit.