Add Labelled Gauges to Metrics System
For rocks DB we need to use gauges to track native stats.  We are also
moving to use multiple databases for the privacy work.  That is easy
to support of we label the databases.

This PR adds labeling to the gauges.  Because gauges do not leave
counters or timers behind this approach returns a consumer that
conusumes the lambda that represents the metric.