[PAN-2954] Support for NoReward and NoProof seal engines
Retesteth requires support for mining without block rewards and without PoW.
No PoW simply requires exposing the nonce generator an dusing a special
EthHashSolver.

No Reward in the tests is interpreted to mean Zero Reward, which presents
problems pre EIP158 with regard to zero balance accounts as the tests point the
coinbase to a non-existent account.  We need to add a sentinel NO_REWARD reward
value (in addition to existing support for a zero value which is a short
circuit) that will proceed with the block reward logic at zero value.
UInt256.MAX_VALUE was chosen as that sentinel as it is a value that would never
make sense otherwise, causing instant integer overflow in any real usecase.

Retesteth will use Wei.ZERO or Wei.NO_REWARD based on the configured fork when
mining in no-reward mode.