[PAN-2946] - changes in core JSON-RPC method to support ReTestEth
Some of the methods need to have a changeable reference to stuff like
blockchainqueries.  For the impacted methods the solution is to wrap them
in a Supplier<> interface.  This includes new constructors for re-used methods.

Also include the debug_accountRangeAt method as it's namespaced as debug.

Also, rename the getter methods to align with standards.