[PAN-3097] Use a BigInteger for Network ID
Interally we use a BigInteger for ChainID, but currently use an int for
NetworkID.  Because the default for NetworkID is the same value as ChainID there
are some chains where this will be very problematic, and there is at least
one other long-living chain outside the Int32 value range.

This is a large changelist because the type is baked fairly deep into some of
the other APIs and very much in the test code testing the type.